% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\name{as_pkgdown}
\alias{as_pkgdown}
\title{Generate pkgdown data structure}
\usage{
as_pkgdown(pkg = ".", override = list())
}
\arguments{
\item{pkg}{Path to package.}

\item{override}{An optional named list used to temporarily override
values in \code{_pkgdown.yml}}
}
\description{
Generate pkgdown data structure
}
