% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdown.R
\docType{package}
\name{dynsite-package}
\alias{dynsite}
\alias{dynsite-package}
\title{dynsite: Package with the good things from pkgdown}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Package with the good things from pkgdown
      but for more than one package!
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dynverse/dynverse.github.io}
  \item Report bugs at \url{https://github.com/dynverse/dynverse.github.io}
}

}
\author{
\strong{Maintainer}: Saelens Wouter \email{wouter.saelens@gmail.com}

}
\keyword{internal}
