% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_references.R
\name{link_references}
\alias{link_references}
\title{Walks through all html files inside folder recursively, and adds}
\usage{
link_references(topics, folder, base_url = "")
}
\description{
Walks through all html files inside folder recursively, and adds
}
